Use scripted resolver in ALL tests
